+++ examples/shared_samples/logical_enclosure.rb
@@ -39,6 +39,7 @@ if variant == 'Synergy'
     name: 'OneViewSDK Test Logical Enclosure',
     forceInstallFirmware: false,
     firmwareBaselineUri: nil
+    initialScopeUris: ['/rest/scopes/a5f8ca3d-2cea-4f82-b880-344572eb7271', '/rest/scopes/e0f6b95a-67a6-4718-b42c-1f7d426b730c']
   }
 
   puts "\nCreating a logical enclosure with the name = '#{options[:name]}'."
@@ -77,6 +78,13 @@ puts "\nFound by name: '#{item2[:name]}'.\n  uri = '#{item2[:uri]}'"
 puts "\nGets a logical enclosure with uri '#{log_encl_uri}'"
 item3 = logical_enclosure_class.find_by(@client, uri: log_encl_uri).first
 puts "Found logical enclosure '#{item3[:uri]}'."
+
+if @client.api_version >= 600
+  # Gets a logical enclosure by scopeUris
+  puts "\nGets a logical enclosure with uri '#{log_encl_uri}'"
+  item4 = logical_enclosure_class.get_all_with_query(@client, {"scopeUris": "/rest/scopes/a5f8ca3d-2cea-4f82-b880-344572eb7271")
+  puts "Found logical enclosure '#{item3[:uri]}'."
+end
 
 puts "\nUpdating a logical enclosure with the name = '#{item3['name']}'."
 old_name = item3['name']