+++ spec/integration/resource/api600/synergy/logical_enclosure/delete_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API500::Synergy::LogicalEnclosure
+klass = OneviewSDK::API600::Synergy::LogicalEnclosure
 RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
-  let(:current_client) { $client_500_synergy }
-  include_examples 'LogicalEnclosureDeleteExample', 'integration api500 context', 'Synergy'
+  let(:current_client) { $client_600_synergy }
+  include_examples 'LogicalEnclosureDeleteExample', 'integration api600 context', 'Synergy'
 end